Añadiendo peticiones en modulo de ventas para crear venta para obtener una venta

+++ src/main/java/cl/duocuc/ecomarket/modelo/dto/venta/DetalleVentaDTO.java
@@ -3,7 +3,7 @@ package cl.duocuc.ecomarket.modelo.dto.venta;
 import java.math.BigDecimal;
 
 public record DetalleVentaDTO (
-        Integer idProducto,
+        Long idProducto,
         Integer cantidad,
         BigDecimal precioUnitario
 ){}
